pyproject.toml
[project]
name = "github-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "pydantic>=2.13.2",
    "pygithub>=2.9.1",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "uvicorn>=0.44.0",
]
